+++ src/Exceptions/tests/Renderer/HighlighterTest.php
@@ -14,7 +14,6 @@ namespace Spiral\Tests\Exceptions\Renderer;
 use PHPUnit\Framework\TestCase;
 use Spiral\Exceptions\Renderer\Highlighter;
 use Spiral\Exceptions\Style\ConsoleStyle;
-use Spiral\Exceptions\Style\HtmlStyle;
 use Spiral\Exceptions\Style\PlainStyle;
 
 class HighlighterTest extends TestCase
@@ -22,20 +21,6 @@ class HighlighterTest extends TestCase
     public function testPlainHighlighter(): void
     {
         $highlighter = new Highlighter(new PlainStyle());
-
-        $this->assertStringContainsString('HighlighterTest', $highlighter->highlight(file_get_contents(__FILE__)));
-    }
-
-    public function testHtmlHighlighter(): void
-    {
-        $highlighter = new Highlighter(new HtmlStyle(HtmlStyle::DEFAULT));
-
-        $this->assertStringContainsString('HighlighterTest', $highlighter->highlight(file_get_contents(__FILE__)));
-    }
-
-    public function testInvertedHtmlHighlighter(): void
-    {
-        $highlighter = new Highlighter(new HtmlStyle(HtmlStyle::INVERTED));
 
         $this->assertStringContainsString('HighlighterTest', $highlighter->highlight(file_get_contents(__FILE__)));
     }
@@ -50,26 +35,6 @@ class HighlighterTest extends TestCase
     public function testPlainHighlighterLines(): void
     {
         $highlighter = new Highlighter(new PlainStyle());
-
-        $this->assertStringContainsString(
-            'HighlighterTest',
-            $highlighter->highlightLines(file_get_contents(__FILE__), 17)
-        );
-    }
-
-    public function testHtmlHighlighterLines(): void
-    {
-        $highlighter = new Highlighter(new HtmlStyle(HtmlStyle::DEFAULT));
-
-        $this->assertStringContainsString(
-            'HighlighterTest',
-            $highlighter->highlightLines(file_get_contents(__FILE__), 17)
-        );
-    }
-
-    public function testInvertedHtmlHighlighterLines(): void
-    {
-        $highlighter = new Highlighter(new HtmlStyle(HtmlStyle::INVERTED));
 
         $this->assertStringContainsString(
             'HighlighterTest',