+++ CHANGELOG.md
@@ -63,6 +63,7 @@
   - [spiral/filters] Added return type `void` and `mixed` parameter type of `$context` to the method `setContext`,
     added return type `mixed` to the method `getContext` in `Spiral\Filters\FilterInterface` interface.
     Added return type `mixed` to the method `getValue` in `Spiral\Filters\InputInterface`.
+  - [spiral/dumper] The `Dumper` Component has been removed from the Framework.
   - [spiral/http] Config `Spiral\Config\JsonPayloadConfig` moved to the `Spiral\Bootloader\Http\JsonPayloadConfig`.
   - [spiral/reactor] Added return type `mixed` and `array|string` parameter type of `$search`,
     `array|string` parameter type of `$replace` to the method `replace` in `Spiral\Reactor\ReplaceableInterface`.